pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "tanx-connector"
version = "0.1.0"
authors = [
  { name="tanX" },
]
license = { file = "LICENSE" }
description = "The official Python connector for Tanx's API"
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["requests", "web3", "ecdsa", "websockets", "sympy", "typing-extensions"]
keywords = ["tanx", "connector", "finance", "api", "wrapper", "python" ]

[project.urls]
"Homepage" = "https://github.com/tanx-libs/tanx-connector-python"
"Bug Tracker" = "https://github.com/tanx-libs/tanx-connector-python/issues"

[project.optional-dependencies]
dev = ["eth-tester", "pytest", "pytest-mock", "responses", "pip-tools", "bumpver", "build", "twine", "python-dotenv"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"tanxconnector.bin" = ["*.json"]

[tool.bumpver]
current_version = "0.1.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "version: v{old_version} -> v{new_version}"
commit = true
tag = true
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']